refactor(): split the app into smaller modules
