+++ package.json
@@ -14,7 +14,8 @@
   "dependencies": {
     "express": "^4.17.1",
     "got": "^11.3.0",
-    "jest": "^26.0.1"
+    "jest": "^26.0.1",
+    "p-map": "^4.0.0"
   },
   "devDependencies": {
     "nodemon": "^2.0.4"
