--- src/helpers.test.js
@@ -1,16 +1,0 @@
-const { formatLanguages } = require("./helpers");
-
-describe("Check formatLanguages", () => {
-  it("should output a single language", () => {
-    const languages = "javascript";
-    expect(formatLanguages).toBeDefined();
-    expect(formatLanguages(languages)).toEqual(`+language:"javascript"`);
-  });
-  it("should output multiple languages", () => {
-    const languages = "java,javascript";
-    expect(formatLanguages).toBeDefined();
-    expect(formatLanguages(languages)).toEqual(
-      `+language:"java"+language:"javascript"`
-    );
-  });
-});
