+++ src/utils.js
@@ -1,0 +1,6 @@
+const formatLanguages = (languageInput = "") => {
+  const languageArray = languageInput.split(",");
+  return languageArray.map((item) => `language:"${item}"`);
+};
+
+module.exports = { formatLanguages };
